Refactor : 파티 목록 로직 분리
커스텀훅을 이용하여 파티 목록을 불러오는 로직을 분리하였습니다.
